Adding support for deleting a file.
Add a lib/aws-s3-delete that has a function to delete the object on
S3.  This function will not return an error - if the file can't be
deleted just keep going.  Modified the controller to delete the record
from the database and to call S3 to delete the object in the bucket.